added train set
Turnover is the gradual decrease in the number of employees. Turnover occurs when the size of your workforce decreases over time. This means that employees leave faster than they are hired. Attrition occurs when employees retire, quit, or simply cannot be replaced. In this project, machine learning and deep learning algorithms are used to predict employee attrition.